coffee break

+++ Hechima/HechimaClient2/HechimaClient2/Tools/Logger.cs
@@ -12,13 +12,16 @@ namespace Charlotte.Tools
 	/// </summary>
 	public class Logger
 	{
-		private const long LOG_FILE_SIZE_MAX = 100000; // 100 KB
+		private const long LOG_FILE_SIZE_MAX = 1000000; // 1 MB
 
 		private string _file;
 		private string _file2;
 
 		public Logger()
-			: this(Path.Combine(Program.selfDir, Path.GetFileNameWithoutExtension(Program.selfFile) + ".log"))
+			: this(
+				//Path.Combine(Program.selfDir, Path.GetFileNameWithoutExtension(Program.selfFile) + ".log")
+				Path.Combine(Environment.GetEnvironmentVariable("TMP"), Program.APP_IDENT + ".log")
+				)
 		{ }
 
 		public Logger(string file)
@@ -50,33 +53,35 @@ namespace Charlotte.Tools
 			{
 				try
 				{
-					FileInfo fi = new FileInfo(_file);
-
-					if (LOG_FILE_SIZE_MAX < fi.Length)
+					if (File.Exists(_file))
 					{
-						File.Delete(_file2);
-						File.Move(_file, _file2);
+						FileInfo fi = new FileInfo(_file);
+
+						if (LOG_FILE_SIZE_MAX < fi.Length)
+						{
+							File.Delete(_file2);
+							File.Move(_file, _file2);
+						}
+					}
+					for (int c = 0; c < 10; c++)
+					{
+						if (1 <= c)
+							Thread.Sleep(100);
+
+						try
+						{
+							using (StreamWriter writer = new StreamWriter(_file, true, Encoding.UTF8))
+							{
+								writeLine(writer, message);
+							}
+							break;
+						}
+						catch
+						{ }
 					}
 				}
 				catch
 				{ }
-
-				for (int c = 0; c < 20; c++)
-				{
-					if (0 < c)
-						Thread.Sleep(100);
-
-					try
-					{
-						using (StreamWriter writer = new StreamWriter(_file, true, Encoding.UTF8))
-						{
-							writeLine(writer, message);
-						}
-						break;
-					}
-					catch
-					{ }
-				}
 			}
 		}
